Do test eigen (I always forget to edit CMakeLists.txt).
We never promised that Lazy_exact_nt would be constructible from unsigned when ET isn't.

+++ Number_types/test/Number_types/Lazy_exact_nt.cpp
@@ -130,7 +130,7 @@ int main ()
   // NT b(a);
   // NT c = b;
   NT d (1.0);
-  NT zzz = 1u;
+  NT zzz = (short)1;
   NT e = d + d;
   NT z = CGAL::my_min(e,d);
   (void) d; (void) e; (void) z; // Shut up warnings.
